+++ hophop/Cargo.toml
@@ -23,6 +23,7 @@ ts-103-636-numbers = { path = "../ts-103-636-numbers/", features = ["defmt"] }
 ts-103-636-utils = { path = "../ts-103-636-utils/", features = ["defmt"] }
 static_cell = "2.1.1"
 itertools = { version = "0.14.0", default-features = false }
+embassy-futures = "0.1.2"
 
 [target.'cfg(not(target = "thumbv8m.main-none-eabihf"))'.dependencies]
 # We're using box_pool, and while that works on thumbv8m without portable
